refactor: reorganize user service imports and add new user data services for improved structure and functionality
