+++ frontend_web/react+vite/src/modules/reports/services/users/getUsersDataService.js
@@ -1,0 +1,22 @@
+import { apiRoutes } from "../../../../config/apiRoutes";
+import { getToken } from "../../../../utils/auth";
+
+export async function getUsersDataService(signal) {
+  const response = await fetch(
+    `${apiRoutes.apiUrl}${apiRoutes.reports}/get_users_by_status`,
+    {
+      method: "GET",
+      headers: {
+        Authorization: getToken(),
+      },
+      signal,
+    },
+  );
+
+  if (!response.ok) {
+    throw new Error("Error en la petición");
+  }
+  const data = await response.json();
+
+  return data.data;
+}
