+++ frontend_web/react+vite/src/modules/reports/services/users/getUsersTableDataService.js
@@ -1,9 +1,9 @@
-import { apiRoutes } from "../../../config/apiRoutes";
-import { getToken } from "../../../utils/auth";
+import { apiRoutes } from "../../../../config/apiRoutes";
+import { getToken } from "../../../../utils/auth";
 
-export async function getUsersPieDataService(signal) {
+export async function getUsersTableDataService(signal) {
   const response = await fetch(
-    `${apiRoutes.apiUrl}${apiRoutes.reports}/get_users_by_rol`,
+    `${apiRoutes.apiUrl}${apiRoutes.reports}/get_users`,
     {
       method: "GET",
       headers: {
